+++ Cargo.toml
@@ -15,11 +15,13 @@ rand = { version = "0.8", default-features = false, optional = true }
 serde = { version = "1.0", default-features = false, features = ["derive"], optional = true }
 
 [dev-dependencies]
-rand = "0.8"
+fuel-types = { path = ".", features = ["random"] }
+rand = { version = "0.8", default-features = false, features = ["std_rng"] }
 
 [features]
 default = ["std"]
-std = []
+alloc = []
 random = ["rand"]
 serde-types = ["serde-types-minimal", "serde/default", "std"]
 serde-types-minimal = ["serde"]
+std = ["alloc"]
